+++ daemon/docs/docs.go
@@ -5198,6 +5198,11 @@ const docTemplate = `{
                         "format": "float64"
                     }
                 },
+                "cpu_power_watts": {
+                    "description": "CPU package power in watts (only present when Intel RAPL is available)",
+                    "type": "number",
+                    "example": 65.5
+                },
                 "cpu_temp_celsius": {
                     "type": "number",
                     "example": 45
@@ -5210,6 +5215,11 @@ const docTemplate = `{
                     "description": "CPU Information",
                     "type": "number",
                     "example": 45.2
+                },
+                "dram_power_watts": {
+                    "description": "DRAM power in watts (only present when Intel RAPL is available)",
+                    "type": "number",
+                    "example": 5.2
                 },
                 "fans": {
                     "description": "Additional Metrics",